backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
bcrypt>=5.0.0
python-jose
python-multipart
pydantic[email]
